+++ coroutines/compose/src/commonMain/kotlin/dev/inmo/micro_utils/coroutines/compose/FlowState.kt
@@ -1,10 +1,7 @@
 package dev.inmo.micro_utils.coroutines.compose
 
 import androidx.compose.runtime.MutableState
-import androidx.compose.runtime.State
-import androidx.compose.runtime.derivedStateOf
 import dev.inmo.micro_utils.coroutines.SpecialMutableStateFlow
-import dev.inmo.micro_utils.coroutines.doInUI
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 
@@ -12,6 +9,7 @@ import kotlinx.coroutines.Dispatchers
  * This type works like [MutableState], [kotlinx.coroutines.flow.StateFlow] and [kotlinx.coroutines.flow.MutableSharedFlow].
  * Based on [SpecialMutableStateFlow]
  */
+@Deprecated("Will be removed soon")
 class FlowState<T>(
     initial: T,
     internalScope: CoroutineScope = CoroutineScope(Dispatchers.Default)
@@ -25,9 +23,9 @@ class FlowState<T>(
             tryEmit(value)
         }
 
-    override suspend fun onChange(value: T) {
+    override fun onChangeWithoutSync(value: T) {
         internalValue = value
-        super.onChange(value)
+        super.onChangeWithoutSync(value)
     }
 
     override fun component1(): T = value