+++ coroutines/src/commonMain/kotlin/dev/inmo/micro_utils/coroutines/SpecialMutableStateFlow.kt
@@ -3,10 +3,14 @@ package dev.inmo.micro_utils.coroutines
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.InternalCoroutinesApi
 import kotlinx.coroutines.channels.BufferOverflow
 import kotlinx.coroutines.flow.FlowCollector
 import kotlinx.coroutines.flow.MutableSharedFlow
+import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.internal.SynchronizedObject
+import kotlinx.coroutines.internal.synchronized
 
 /**
  * Works like [StateFlow], but guarantee that latest value update will always be delivered to
@@ -15,7 +19,9 @@ import kotlinx.coroutines.flow.StateFlow
 open class SpecialMutableStateFlow<T>(
     initialValue: T,
     internalScope: CoroutineScope = CoroutineScope(Dispatchers.Default)
-) : StateFlow<T>, FlowCollector<T>, MutableSharedFlow<T> {
+) : MutableStateFlow<T>, FlowCollector<T>, MutableSharedFlow<T> {
+    @OptIn(InternalCoroutinesApi::class)
+    private val syncObject = SynchronizedObject()
     protected val internalSharedFlow: MutableSharedFlow<T> = MutableSharedFlow(
         replay = 0,
         extraBufferCapacity = 2,
@@ -28,22 +34,42 @@ open class SpecialMutableStateFlow<T>(
     )
 
     protected var _value: T = initialValue
-    override val value: T
+    override var value: T
         get() = _value
-    protected open suspend fun onChange(value: T) {
-        _value = value
-        publicSharedFlow.emit(value)
-    }
-    protected val job = internalSharedFlow.subscribe(internalScope) {
-        if (_value != it) {
-            onChange(it)
+        set(value) {
+            doOnChangeAction(value)
         }
+    protected val job = internalSharedFlow.subscribe(internalScope) {
+        doOnChangeAction(it)
     }
 
     override val replayCache: List<T>
         get() = publicSharedFlow.replayCache
     override val subscriptionCount: StateFlow<Int>
         get() = publicSharedFlow.subscriptionCount
+
+    @OptIn(InternalCoroutinesApi::class)
+    override fun compareAndSet(expect: T, update: T): Boolean {
+        return synchronized(syncObject) {
+            if (expect == _value && update != _value) {
+                doOnChangeAction(update)
+            }
+            expect == _value
+        }
+    }
+
+    protected open fun onChangeWithoutSync(value: T) {
+        _value = value
+        publicSharedFlow.tryEmit(value)
+    }
+    @OptIn(InternalCoroutinesApi::class)
+    protected open fun doOnChangeAction(value: T) {
+        synchronized(syncObject) {
+            if (_value != value) {
+                onChangeWithoutSync(value)
+            }
+        }
+    }
 
     @ExperimentalCoroutinesApi
     override fun resetReplayCache() = publicSharedFlow.resetReplayCache()