+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.20.18
+
+* `Coroutines`:
+    * `SpecialMutableStateFlow` now extends `MutableStateFlow`
+    * `Compose`:
+        * Deprecate `FlowState` due to its complexity in fixes
+
 ## 0.20.17
 
 * `Versions`: